RAG/requirements.txt
# For MLFlow evaluation
mlflow==3.3.2
pandas==2.3.2
evaluate==0.4.5
textstat==0.7.8
tiktoken==0.11.0
torch==2.8.0
tokenizers==0.21.4
transformers==4.55.4
sentence-transformers==4.1.0
openpyxl==3.1.5
python-dotenv==1.1.1
# For RAG
chromadb==1.0.15
fastapi[standard]==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
langchain==1.0.3
langchain-chroma==0.2.5
langchain-community==0.4.1
langchain-openai==1.0.1
langgraph==1.0.2
openai==1.109.1
# For Ragas evaluation
ragas==0.3.8